BlockKit用関数Labelの内包化 (#119)
* update: sectionArgsをinterfaceからtypeに変更

* update: labelを各blockKit用関数に内包するようにした

* update: generalComponent下の関数の定義を明記

* change: clubInfoをgetMessageにファイル名変更

* change: modal部分をフォルダ/ファイル分け

* chore: 🤖 Windows でも動くように

* fix: 🐛 functions/type's name & change I/F name upper camel
