+++ package.json
@@ -8,8 +8,8 @@
   },
   "scripts": {
     "build": "./node_modules/.bin/tsc",
-    "lint": "eslint --ext './src/**/*.ts' './**/*.json' --cache",
-    "lint:fix": "eslint --fix --ext './src/**/*.ts' './**/*.json' --cache",
+    "lint": "eslint --ext \"./src/**/*.ts\" \"./**/*.json\" --cache",
+    "lint:fix": "eslint --fix --ext \"./src/**/*.ts\" \"./**/*.json\" --cache",
     "firebase:serve": "firebase emulators:start --only functions",
     "firebase:shell": "firebase functions:shell",
     "firebase:start": "npm run shell",
